feat: Enhance API documentation and delete functionality across multiple controllers
- Added Swagger annotations to improve API documentation for EspacioFavoritoController, EstacionamientoController, EstadoEstacionamientoController, EstadoIncidenteController, EstadoReservaController, FeedbackController, IncidenteController, ReservaController, SensorController, TipoEstacionamientoController, TipoEstadoIncidenteController, TipoEstadoReservaController.
- Implemented delete methods with appropriate response handling in several controllers.
- Updated service interfaces and implementations to support delete operations with boolean return types for better error handling.

+++ gestion-estacionamiento/src/main/java/cl/alcoholicos/gestorestacionamiento/service/impl/FeedbackService.java
@@ -47,9 +47,12 @@ public class FeedbackService implements IFeedback {
     }
 
     @Override
-    public FeedbackEntity delete(Integer idFeedback) {
-        feedbackRepository.deleteById(idFeedback);
-        return null;
+    public boolean delete(Integer idFeedback) {
+        if (feedbackRepository.existsById(idFeedback)) {
+            feedbackRepository.deleteById(idFeedback);
+            return true;
+        }
+        return false;
     }
 
     @Override
